requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3
langgraph==0.2.24 
streamlit==1.39.0 
python-dotenv==1.0.1 
requests==2.32.3 
nest-asyncio==1.6.0 
httpx==0.27.2 
fastapi==0.115.0 
uvicorn==0.30.6